#15 created a dialog containing the TagAndTagGroupEditor
The TagAndTagGroupEditorDialog contains an instance of the
TagAndTagGroupEditor and two buttons to save the changes made in the
editor and to cancel the dialog.

The dialog is shown after a click on the new button in the
AdditionalTagsPanel and the newly created tags in the editor will be
shown in a list in the AdditionalTagsPanel.