#15 added getPersistedCidsBeans() to EditorShowableInDialog
The method getPersistedCidsBeans() was added to the interface
EditorShowableInDialog. This had to be done because if a cidsBean was
persisted in the dialog, the persisted cidsBean has to be replaced with
its original in the caller of the dialog. If this is not done the
persisted bean will be saved twice.