#41 created a non modal version of ShowEditorInDialog
A non modal version of ShowEditorInDialog is needed, which behaves
differently from its parent class, because in the non-modal case
showDialog() returns immediately. Therefore the non-modal version needs
to inform its listeners if it changes were saved successfully. This is
not needed in the modal version as the program runs sequentially.