#80 created the inner class StandardMetaDataFetcher
The main task of StandardMetaDataFetcher is to fetch the standard meta data from the
database and add it to a resource. To achieve this it tries to fetch the
standard meta data-CidsBean and a Switch-On contact-CidsBean. If this
fails it creates a new CidsBean, using the methods from
DefaultPropertySetter. Finally the standard meta data-CidsBean is added
to the resource-CidsBean.