#92 delay the first appearance of a warning by 500 ms
A small hack is needed to avoid that the warning is
shown to early. After the settings were read, the warning can not be
shown for a half second. This has to be done as the warning will be
shown when the first component, which contains a mandatory value,
receives the focus. Due to the binding on the components the
isValid()-method will be fired.