Fixes #102: check for WizardDescriptor
This bug was introduced in 7382bf6. In that commit the
PropertyChangeListener of the wizard, checked if the wizard was finished
with evt.getNewValue() == FINISH_OPTION. The problem is that such an
event is fired twice.
Therefore it has to be checked if evt.getNewValue() is an instance of
the class WizardDescriptor.FinishAction. Although the keyword instanceof
can not be used as that class is an inner class of WizardDescriptor an
has the default access modifier.