[issue #119] configurable webdav
The webdav configuration for data import is not available from the jar's
resources anymore. It has been moved to the res.jar where it can be
configured by any admin. Moreover the password is not in clear text
anymore but has to be provided in an encrypted form that is known to the
cismet PasswordEncrypter. Finally, the mime type recognition is not
working properly everytime so the system now handles unknown mime types
without crasing.